// src/redux/actions/types/schedulesActionTypes.ts
export enum ESchedulesActionTypes {
  CONFIRM_DELETE = 'SCHEDULES/CONFIRM_DELETE',
  LOADING = 'SCHEDULES/LOADING',
  LOADING_RECORDS = 'SCHEDULES/LOADING_RECORDS',
  LOADING_CALLING_LIST = 'SCHEDULES/LOADING_CALLING_LIST',
  SET = 'SCHEDULES/SET',
  ADD_RECORDS = 'SCHEDULES/ADD_RECORDS',
  ADD_SCHEDULE = 'SCHEDULES/ADD_SCHEDULE',
  DELETE_SCHEDULE = 'SCHEDULES/DELETE_SCHEDULE',
  ADD_RECORDS_SINGLE = 'SCHEDULES/ADD_RECORDS_SINGLE',
  ADD_ITEMS = 'SCHEDULES/ADD_ITEMS',
  ADD_ACTIVE = 'SCHEDULES/ADD_ACTIVE',
  ADD_ACTIVE_FOLDER = 'SCHEDULES/ADD_ACTIVE_FOLDER',
  TOGLE_EXPANDE = 'SCHEDULES/TOGLE_EXPANDE',
  TOGLE_LOADING = 'SCHEDULES/TOGLE_LOADING',
  SET_POPUP = 'SCHEDULES/SET_POPUP',
  SET_EDIT_POPUP = 'SCHEDULES/SET_EDIT_POPUP',
  SET_EDITAll_POPUP = 'SCHEDULES/SET_EDITAll_POPUP',
  ADD_FIELDS = 'SCHEDULES/ADD_FIELDS',
  TOGLE_CHECKED_RECCORD = 'SCHEDULES/TOGLE_CHECKED_RECCORD',
  TOGLE_CHECKED_ALL = 'SCHEDULES/TOGLE_CHECKED_ALL',
  SET_TOTAL_COUNT = 'SCHEDULES/SET_TOTAL_COUNT',
  CHENGE_PARAM = 'SCHEDULES/CHENGE_PARAM',
  UPDATE_COLORS = 'SCHEDULES/UPDATE_COLORS',
  UPDATE_NAME = 'SCHEDULES/UPDATE_NAME',
  USE_FILTER = 'SCHEDULES/USE_FILTER',
  SET_FILTER = 'SCHEDULES/SET_FILTER',
  CHANGE_ORDER_BY_CONDITION = 'SCHEDULES/CANGE_ORDER_BY',
  CHANGE_WHERE_CONDITION = 'SCHEDULES/CANGE_WHERE_CONDITION',
  CAMPAIGN_GROUPS_SCHEDULE = 'SCHEDULES/CAMPAIGN_GROUPS_SCHEDULE',
  CAMPAIGN_GROUPS_SERVERS = 'SCHEDULES/CAMPAIGN_GROUPS_SERVERS',
  SET_INCLUDED = 'SCHEDULES/SET_INCLUDED',
  SET_SHOW_CUSTOM_FIELDS = 'SCHEDULES/SET_SHOW_CUSTOM_FIELDS',

}